# Order of magnitude faster HWCD implementation for trees