Two functions are added here: `hardwiredClusterDistance_treelike` and `hardwiredClusters_treelike`. The runtimes of these functions grow linearly with respect to the number of edges in the input tree whereas the current implementation appears to grow exponentially.

This implementation is most useful when computing unrooted HWCD for large trees, because so many rooting combinations need to be checked.

---

Below are some figures showing that this implementation is marginally slower than the current implementation with fewer than ~75 taxa, but scales significantly better.

**Runtime:**

<img src="https://github.com/user-attachments/assets/6f5a9bd5-482d-4afe-8960-636779408f80" width="48%"/>
<img src="https://github.com/user-attachments/assets/2f26ae54-7bad-474d-87a2-b5e527216f46" width="48%"/>

At `n=1000`, this new implementation is ~152x faster (27ms vs. 4.1s).

**Memory Usage:**

<img src="https://github.com/user-attachments/assets/78791079-e11e-4b9c-9fd8-511aea46d641" width="48%"/>
<img src="https://github.com/user-attachments/assets/bbf47f54-64c0-42cb-9ede-7dffec868c2a" width="48%"/>

At `n=1000`, this new implementation uses ~1122x less memory (10.4 MiB vs. 11.4 GiB).

---

Below are benchmarks where `tre` is a 1,000 taxa tree that can be found [here](https://gist.githubusercontent.com/NathanKolbow/252b12c5769eadb7dc8337c7005c4387/raw/fc3b9197104df90702fa9b54252016f5aa903ec1/my.tre).

**Old implementation:**

```julia
julia> @benchmark hardwiredClusterDistance(tre, tre, true)
BenchmarkTools.Trial: 2 samples with 1 evaluation.
 Range (min … max):  3.585 s …   3.632 s  ┊ GC (min … max): 8.82% … 8.74%
 Time  (median):     3.609 s              ┊ GC (median):    8.78%
 Time  (mean ± σ):   3.609 s ± 32.832 ms  ┊ GC (mean ± σ):  8.78% ± 0.06%

  █                                                                   █
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  3.59 s         Histogram: frequency by time        3.63 s <

 Memory estimate: 11.37 GiB, allocs estimate: 1502468.
```

**New implementation:**

```julia
julia> @benchmark hardwiredClusterDistance(tre, tre, true)
BenchmarkTools.Trial: 203 samples with 1 evaluation.
 Range (min … max):  23.405 ms … 31.016 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     23.956 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   24.690 ms ±  1.430 ms  ┊ GC (mean ± σ):  2.16% ± 3.65%

    █ ▂▂
  ▄█████▅▇▄▄▅▃▁▂▂▁▂▂▁▁▂▁▂▃▁▂▅▅▃▅▄▄▂▄▃▃▂▂▂▁▂▁▁▁▁▁▁▁▁▂▁▁▁▁▂▁▂▁▄ ▃
  23.4 ms         Histogram: frequency by time        28.8 ms <

 Memory estimate: 10.37 MiB, allocs estimate: 36755.
```

---

As written, these functions only work with trees, but they could be adapted to work with networks relatively easily.

**Important note:** `hardwiredClusters_treelike` and `hardwiredClusterDistance_treelike` were written for exclusively internal use, so (1) their docstrings are not very verbose, and (2) the return type of `hardwiredClusters_treelike` **does not match** the return type of `hardwiredClusters`. Both of these points could be changed with relatively little work if these functions are desired over the current implementations and are adapted to work on networks as well.