@cecileane 2024-09-25
This is fantastic! I'm very excited about the gain in speed. I suspect that a similar gain could also be achieved for non-tree networks. 3 things:

1. For now we aren't making new changes to the package, as we are doing major work to shrink `PhyloNetworks` (on branch [`dev`](https://github.com/JuliaPhylo/PhyloNetworks.jl/tree/dev)), and to remove many functionalities from it, to place them in new packages (all in development branches for now): trait evolution in [`PhyloTraits`](https://github.com/JuliaPhylo/PhyloTraits.jl/tree/dev01), [`PhyLiNC`](https://github.com/JuliaPhylo/PhyLiNC.jl/tree/dev01) and [`SNaQ.jl`](https://github.com/JuliaPhylo/SNaQ.jl). If you could help migrate the `snaq!` functions to SNaQ.jl, that would be awesome, because there's just a basic skeleton for now, not even a dev branch. @crsl4 is in charge of this part.
2. It would be fantastic if similar ideas were used for non-tree networks, both for a speed up on all types of networks, but also for a more consistent code base.
3. After the refactoring of PhyloNetworks & co. packages is done, would you be interested in helping to implement the edge-based μ-distance between semidirected networks described [here](https://arxiv.org/abs/2405.16035)?
   * The μ-representation of a network contains more information than the set of hardwired clusters, and so it is a distance on a larger class of networks than the hardwired-cluster distance.
   * On trees, it extends both the rooted RF and the unrooted RF distance (like the hardwired cluster distance).
   * And importantly, it can be computed fast, with *no* need to try all possible rootings of each semidirected network being compared --hence another speed gain.