@cecileane 2024-09-26
> > 3\. After the refactoring of PhyloNetworks & co. packages is done, would you be interested in helping to implement the edge-based μ-distance between semidirected networks described [here](https://arxiv.org/abs/2405.16035)?
> >
> 
> This looks interesting! Does a project/codebase already exist for this? There does not seem to be any software implementation referenced in the paper.

No, but it could be very useful to many. Hence my question!